Made changing the map being used more convenient
Modified the makefile so that when running the game with valgrind using "make v" or "make vb" it is possible to add "MAP=<mapname>" to the end to change what map is being used in the run. e.g. "make vb MAP=ok4" will use ./maps/valid/ok4.ber when running the game